# Reduce allocations in CSharp command line parsing