… src/Compilers/CSharp/Test/CommandLine/CommandLineTests.cs
… src/Compilers/Core/Portable/InternalUtilities/FileNameUtilities.cs
… src/Compilers/Core/Portable/MemoryExtensions.cs
… src/Compilers/VisualBasic/Portable/CommandLine/VisualBasicCommandLineParser.vb
… src/Tools/BuildBoss/CompilerNuGetCheckerUtil.cs