Context #53570

The C# parser is used by the project system in scenarios like solution
open. That means the allocations in command line parsing can contribute
significantly to Visual Studio performance. This PR reduces the
allocations significantly.

Note: in the below explanations when I refer to the "99% case" I am
referring to how command lines are structured when created by
MSBuild based builds. That is the **overwhelming** case for the compiler
and is advantageous because it normalizes many items like lower casing
all option names, having one reference per option, etc ...  The remaining
1% are hand authored build files and these should not impact Visual
Studio scenarios.

The first fix is to simply avoid iterators in the parse hot paths. Most
of the iterators in the parser return a single element in the 99% case
hence the iterator is wasted allocations. These were switched to take
pooled builder arguments.

The next, and more siginificant fix, is to use `ReadOnlyMemory<char>`
instead of `string` on our hot parsing paths. This allows us to parse
the command line arguments without actually allocating memory for the
values until actually needed by lower level APIs. This allowed for APIs
like `RemoveQuotesAndSlashes` to become allocation free in the 99% case
(it's just a slicing operation now, not a string allocation).

The hot paths in our parsing were updated to employ these techniques.

The one downside of the change is that I had to touch virtually every
`case "someOption"` where the option value was used. The value is now
held in a `ReadOnlyMemory<char>` until it's actually needed as a
`string`. Hence every one of these cases either needed to use the new
`ReadOnlyMemory<char>` APIs or force the allocation of the `string`.
That made the change longer than I would've liked but it's also fairly
mechanical.

To test out the changes I created a benchmark which parsed a couple of
command lines:

- Command line for building Microsoft.CodeAnalysis.CSharp
- Command line for a simple razor app

The results are below and represent a significant win in both
allocations and in simple performance. There are a few more places we
could get some wins but getting into the realm of diminishing returns at
this point.

Before

|                          Method |       Mean |    Error |   StdDev |    Gen 0 |    Gen 1 | Gen 2 | Allocated |
|-------------------------------- |-----------:|---------:|---------:|---------:|---------:|------:|----------:|
|  ParseCSharpCompilerCommandLine | 1,101.9 us | 16.05 us | 14.23 us | 242.1875 |  80.0781 |     - |    876 KB |
| ParseCSharpCompilerResponseFile | 1,916.7 us | 18.26 us | 14.26 us | 353.5156 | 175.7813 |     - |  2,082 KB |
|   ParseRazorCompilerCommandLine |   467.3 us |  2.27 us |  2.02 us | 350.5859 |  53.7109 |     - |    571 KB |
|  ParseRazorCompilerResponseFile |   850.6 us |  8.09 us |  7.17 us | 327.1484 | 102.5391 |     - |    953 KB |


After
|                          Method |       Mean |    Error |   StdDev |    Gen 0 |   Gen 1 |  Gen 2 | Allocated |
|-------------------------------- |-----------:|---------:|---------:|---------:|--------:|-------:|----------:|
|  ParseCSharpCompilerCommandLine |   833.7 us | 16.53 us | 31.04 us | 111.3281 | 18.5547 |      - |    200 KB |
| ParseCSharpCompilerResponseFile | 1,631.5 us | 11.42 us | 10.69 us | 167.9688 | 78.1250 | 1.9531 |    891 KB |
|   ParseRazorCompilerCommandLine |   257.6 us |  3.44 us |  3.05 us |  80.0781 | 14.6484 |      - |    136 KB |
|  ParseRazorCompilerResponseFile |   600.0 us |  5.59 us |  5.23 us | 119.1406 | 33.2031 |      - |    364 KB |

There is a Gen2 increase here but I believe that is just the new usage of pooled builders.

Closes #53570