@jaredpar 2021-07-08
@dotnet/roslyn-compiler PTAL