@jaredpar 2021-07-09
After reading through a few of the comments I did some more profiling and found I could save a lot more allocations by avoiding the `TextReader.ReadLine` calls. Under the hood that uses a lot of `StringBuilder` and there was really no reason for that. The majority of lines can be parsed into a stack based `Span<char>` and parsed out directly from there. That produced a significant allocation savings

|                          Method |       Mean |    Error |   StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------------------------- |-----------:|---------:|---------:|---------:|--------:|------:|----------:|
|  ParseCSharpCompilerCommandLine |   771.7 us | 15.09 us | 19.62 us | 109.3750 | 20.5078 |     - |    200 KB |
| ParseCSharpCompilerResponseFile | 2,094.4 us | 20.04 us | 18.74 us | 105.4688 | 39.0625 |     - |    551 KB |
|   ParseRazorCompilerCommandLine |   249.5 us |  2.49 us |  2.33 us |  77.6367 | 14.6484 |     - |    136 KB |
|  ParseRazorCompilerResponseFile |   710.1 us |  5.97 us |  5.29 us |  88.8672 | 24.4141 |     - |    251 KB |